requirements.txt
langchain
langchain-openai
langchain-google-genai
streamlit
streamlit-chat
python-dotenv
langchain-core